Create owm component for SharedEntitiesOverview
